整合了毛新业务

+++ src/main/java/com/neusoft/ssm/bean/RegistrationInfo.java
@@ -9,7 +9,7 @@ public class RegistrationInfo {
   private String medical_record_no;
   private String patient_name;
   private String gender;
-  private Long age;
+  private Double age;
   private Date birthday;
   private String registration_category;
   private String medical_category;
@@ -34,101 +34,101 @@ public class RegistrationInfo {
   private String birthday_date;
   private String settle_category;
 
-    public Long getId() {
-        return id;
-    }
+  public Long getId() {
+    return id;
+  }
 
-    public void setId(Long id) {
-        this.id = id;
-    }
+  public void setId(Long id) {
+    this.id = id;
+  }
 
-    public String getMedical_record_no() {
-        return medical_record_no;
-    }
+  public String getMedical_record_no() {
+    return medical_record_no;
+  }
 
-    public void setMedical_record_no(String medical_record_no) {
-        this.medical_record_no = medical_record_no;
-    }
+  public void setMedical_record_no(String medical_record_no) {
+    this.medical_record_no = medical_record_no;
+  }
 
-    public String getPatient_name() {
-        return patient_name;
-    }
+  public String getPatient_name() {
+    return patient_name;
+  }
 
-    public void setPatient_name(String patient_name) {
-        this.patient_name = patient_name;
-    }
+  public void setPatient_name(String patient_name) {
+    this.patient_name = patient_name;
+  }
 
-    public String getGender() {
-        return gender;
-    }
+  public String getGender() {
+    return gender;
+  }
 
-    public void setGender(String gender) {
-        this.gender = gender;
-    }
+  public void setGender(String gender) {
+    this.gender = gender;
+  }
 
-    public Long getAge() {
-        return age;
-    }
+  public Double getAge() {
+    return age;
+  }
 
-    public void setAge(Long age) {
-        this.age = age;
-    }
+  public void setAge(Double age) {
+    this.age = age;
+  }
 
-    public Date getBirthday() {
-        return birthday;
-    }
+  public Date getBirthday() {
+    return birthday;
+  }
 
-    public void setBirthday(Date birthday) {
-        this.birthday = birthday;
-    }
+  public void setBirthday(Date birthday) {
+    this.birthday = birthday;
+  }
 
-    public String getRegistration_category() {
-        return registration_category;
-    }
+  public String getRegistration_category() {
+    return registration_category;
+  }
 
-    public void setRegistration_category(String registration_category) {
-        this.registration_category = registration_category;
-    }
+  public void setRegistration_category(String registration_category) {
+    this.registration_category = registration_category;
+  }
 
-    public String getMedical_category() {
-        return medical_category;
-    }
+  public String getMedical_category() {
+    return medical_category;
+  }
 
-    public void setMedical_category(String medical_category) {
-        this.medical_category = medical_category;
-    }
+  public void setMedical_category(String medical_category) {
+    this.medical_category = medical_category;
+  }
 
-    public String getIdentity_card_no() {
-        return identity_card_no;
-    }
+  public String getIdentity_card_no() {
+    return identity_card_no;
+  }
 
-    public void setIdentity_card_no(String identity_card_no) {
-        this.identity_card_no = identity_card_no;
-    }
+  public void setIdentity_card_no(String identity_card_no) {
+    this.identity_card_no = identity_card_no;
+  }
 
-    public String getFamily_address() {
-        return family_address;
-    }
+  public String getFamily_address() {
+    return family_address;
+  }
 
-    public void setFamily_address(String family_address) {
-        this.family_address = family_address;
-    }
+  public void setFamily_address(String family_address) {
+    this.family_address = family_address;
+  }
 
-    public Date getRegistration_date() {
-        return registration_date;
-    }
+  public Date getRegistration_date() {
+    return registration_date;
+  }
 
-    public void setRegistration_date(Date registration_date) {
-        this.registration_date = registration_date;
-    }
+  public void setRegistration_date(Date registration_date) {
+    this.registration_date = registration_date;
+  }
 
-    public Date getSee_doctor_date() {
-        return see_doctor_date;
-    }
+  public Date getSee_doctor_date() {
+    return see_doctor_date;
+  }
 
-    public void setSee_doctor_date(Date see_doctor_date) {
-        this.see_doctor_date = see_doctor_date;
-    }
+  public void setSee_doctor_date(Date see_doctor_date) {
+    this.see_doctor_date = see_doctor_date;
+  }
 
   public String getDepartment_id() {
     return department_id;
@@ -138,21 +138,21 @@ public class RegistrationInfo {
     this.department_id = department_id;
   }
 
-    public Long getDoctor_id() {
-        return doctor_id;
-    }
+  public Long getDoctor_id() {
+    return doctor_id;
+  }
 
-    public void setDoctor_id(Long doctor_id) {
-        this.doctor_id = doctor_id;
-    }
+  public void setDoctor_id(Long doctor_id) {
+    this.doctor_id = doctor_id;
+  }
 
-    public String getRegistration_source() {
-        return registration_source;
-    }
+  public String getRegistration_source() {
+    return registration_source;
+  }
 
-    public void setRegistration_source(String registration_source) {
-        this.registration_source = registration_source;
-    }
+  public void setRegistration_source(String registration_source) {
+    this.registration_source = registration_source;
+  }
 
   public Integer getSettle_accounts_category() {
     return settle_accounts_category;
@@ -162,21 +162,21 @@ public class RegistrationInfo {
     this.settle_accounts_category = settle_accounts_category;
   }
 
-    public String getIs_seen_doctor() {
-        return is_seen_doctor;
-    }
+  public String getIs_seen_doctor() {
+    return is_seen_doctor;
+  }
 
-    public void setIs_seen_doctor(String is_seen_doctor) {
-        this.is_seen_doctor = is_seen_doctor;
-    }
+  public void setIs_seen_doctor(String is_seen_doctor) {
+    this.is_seen_doctor = is_seen_doctor;
+  }
 
-    public String getStatus() {
-        return status;
-    }
+  public String getStatus() {
+    return status;
+  }
 
-    public void setStatus(String status) {
-        this.status = status;
-    }
+  public void setStatus(String status) {
+    this.status = status;
+  }
 
   public Double getExpense() {
     return expense;
@@ -186,61 +186,61 @@ public class RegistrationInfo {
     this.expense = expense;
   }
 
-    public String getReserve1() {
-        return reserve1;
-    }
+  public String getReserve1() {
+    return reserve1;
+  }
 
-    public void setReserve1(String reserve1) {
-        this.reserve1 = reserve1;
-    }
+  public void setReserve1(String reserve1) {
+    this.reserve1 = reserve1;
+  }
 
-    public String getReserve2() {
-        return reserve2;
-    }
+  public String getReserve2() {
+    return reserve2;
+  }
 
-    public void setReserve2(String reserve2) {
-        this.reserve2 = reserve2;
-    }
+  public void setReserve2(String reserve2) {
+    this.reserve2 = reserve2;
+  }
 
-    public String getReserve3() {
-        return reserve3;
-    }
+  public String getReserve3() {
+    return reserve3;
+  }
 
-    public void setReserve3(String reserve3) {
-        this.reserve3 = reserve3;
-    }
+  public void setReserve3(String reserve3) {
+    this.reserve3 = reserve3;
+  }
 
-    public String getDoctor_name() {
-        return doctor_name;
-    }
+  public String getDoctor_name() {
+    return doctor_name;
+  }
 
-    public void setDoctor_name(String doctor_name) {
-        this.doctor_name = doctor_name;
-    }
+  public void setDoctor_name(String doctor_name) {
+    this.doctor_name = doctor_name;
+  }
 
-    public String getDepartment_name() {
-        return department_name;
-    }
+  public String getDepartment_name() {
+    return department_name;
+  }
 
-    public void setDepartment_name(String department_name) {
-        this.department_name = department_name;
-    }
+  public void setDepartment_name(String department_name) {
+    this.department_name = department_name;
+  }
 
-    public String getDate() {
-        return date;
-    }
+  public String getDate() {
+    return date;
+  }
 
-    public void setDate(String date) {
-        this.date = date;
-    }
+  public void setDate(String date) {
+    this.date = date;
+  }
 
-    public String getBirthday_date() {
-        return birthday_date;
-    }
+  public String getBirthday_date() {
+    return birthday_date;
+  }
 
-    public void setBirthday_date(String birthday_date) {
-        this.birthday_date = birthday_date;
-    }
+  public void setBirthday_date(String birthday_date) {
+    this.birthday_date = birthday_date;
+  }
 
   public String getSettle_category() {
     return settle_category;
